requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
pytest>=7.0.0
ruff>=0.1.0